@@ -1,13 +1,13 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 28
-    buildToolsVersion '28.0.3'
+    compileSdkVersion 32
+    buildToolsVersion '30.0.3'
 
     defaultConfig {
         applicationId "com.livejournal.karino2.voiceexpense"
-        minSdkVersion 19
-        targetSdkVersion 28
+        minSdk 24
+        targetSdkVersion 32
         versionCode 11
         versionName "1.1"
     }
@@ -21,6 +21,5 @@ android {
 
 dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
-    implementation 'com.android.support:appcompat-v7:28.0.0'
-    // implementation 'com.google.android.gms:play-services:12.0.1'
+    implementation 'androidx.appcompat:appcompat:1.4.1'
 }
